Auth

+++ src/controllers/note_controller.js
@@ -5,20 +5,21 @@ import Person from '../models/person_model';
 // Problem to fix later: Can't unassociate note from person
 // Problem to fix later: Can't unassociate note from company
 
-export async function createNote(noteFields) {
+export async function createNote(noteFields, userId) {
   const note = new Note();
   note.title = noteFields.title;
   note.content = noteFields.content || '';
   note.tags = noteFields.tags || [];
   note.associatedCompany = noteFields.associatedCompany || null;
   note.associatedPerson = noteFields.associatedPerson || null;
+  note.author = userId;
 
   try {
     await note.validate();
     if (note.associatedPerson) {
-      await addToAssociatedPerson(note.associatedPerson, noteFields.associatedCompany, note);
+      await addToAssociatedPerson(note.associatedPerson, noteFields.associatedCompany, note, userId);
     } else if (note.associatedCompany) {
-      await addToAssociatedCompany(note.associatedCompany, note);
+      await addToAssociatedCompany(note.associatedCompany, note, userId);
     }
     const savedNote = await note.save();
     console.log(savedNote);
@@ -28,29 +29,33 @@ export async function createNote(noteFields) {
   }
 }
 
-export async function getNotes() {
+export async function getNotes(userId) {
   try {
-    const notes = await Note.find({}, 'title tags content');
+    const notes = await Note.find({ author: userId }, 'title tags content');
     return notes;
   } catch (error) {
     throw new Error(`get note error: ${error}`);
   }
 }
 
-export async function findNotes(query) {
+export async function findNotes(query, userId) {
   try {
-    const searchedNote = await Note.find({ $text: { $search: query } }, 'title tags content');
+    const searchedNote = await Note.find({ author: userId, $text: { $search: query } }, 'title tags content');
     return searchedNote;
   } catch (error) {
     throw new Error(`get note error: ${error}`);
   }
 }
 
-export async function getNote(id) {
+export async function getNote(id, userId) {
   try {
     const note = await Note.findById(id);
     if (!note) {
       throw new Error('unable to find note');
+    }
+
+    if (userId !== note.author.toString()) {
+      throw new Error('No permission error');
     }
     return note;
   } catch (error) {
@@ -58,14 +63,21 @@ export async function getNote(id) {
   }
 }
 
-export async function deleteNote(id) {
+export async function deleteNote(id, userId) {
   try {
     const note = await Note.findById(id);
+    if (!note) {
+      throw new Error('unable to find note');
+    }
+
+    if (userId !== note.author.toString()) {
+      throw new Error('No permission error');
+    }
     if (note.associatedPerson) {
-      deleteFromExAssociatedPerson(note);
+      deleteFromExAssociatedPerson(note, userId);
     }
     if (note.associatedCompany) {
-      deleteFromExAssociatedCompany(note);
+      deleteFromExAssociatedCompany(note, userId);
     }
     const deletedNote = await Note.deleteOne({ _id: id });
     return deletedNote;
@@ -74,9 +86,16 @@ export async function deleteNote(id) {
   }
 }
 
-export async function updateNote(id, noteFields) {
+export async function updateNote(id, noteFields, userId) {
   try {
     const note = await Note.findById(id);
+    if (!note) {
+      throw new Error('unable to find note');
+    }
+
+    if (userId !== note.author.toString()) {
+      throw new Error('No permission error');
+    }
     const {
       title, content, tags, associatedCompany, associatedPerson,
     } = noteFields;
@@ -91,17 +110,17 @@ export async function updateNote(id, noteFields) {
     }
     if (associatedPerson && associatedPerson !== note.associatedPerson.toString()) {
       await note.validate();
-      await deleteFromExAssociatedCompany(note);
-      await deleteFromExAssociatedPerson(note);
-      await addToAssociatedPerson(associatedPerson, associatedCompany, note);
+      await deleteFromExAssociatedCompany(note, userId);
+      await deleteFromExAssociatedPerson(note, userId);
+      await addToAssociatedPerson(associatedPerson, associatedCompany, note, userId);
       note.associatedPerson = associatedPerson;
     } else if (associatedCompany && note.associatedCompany.toString() !== associatedCompany) {
       await note.validate();
       if (note.associatedPerson) {
         throw new Error('cannot associate note to a new company if it is already associated with a person in existing company');
       }
-      await deleteFromExAssociatedCompany(note);
-      await addToAssociatedCompany(associatedCompany, note);
+      await deleteFromExAssociatedCompany(note, userId);
+      await addToAssociatedCompany(associatedCompany, note, userId);
       note.associatedCompany = associatedCompany;
     }
     const savedNote = await note.save();
@@ -111,7 +130,7 @@ export async function updateNote(id, noteFields) {
   }
 }
 
-async function addToAssociatedCompany(companyId, note) {
+async function addToAssociatedCompany(companyId, note, userId) {
   try {
     const company = await Company.findById(companyId);
     if (!company) {
@@ -125,9 +144,16 @@ async function addToAssociatedCompany(companyId, note) {
   }
 }
 
-async function deleteFromExAssociatedCompany(note) {
+async function deleteFromExAssociatedCompany(note, userId) {
   try {
     const company = await Company.findById(note.associatedCompany);
+    if (!company) {
+      throw new Error('unable to find company');
+    }
+
+    if (userId !== company.author.toString()) {
+      throw new Error('No permission error');
+    }
     company.notes.pull(note.id);
     const savedCompany = await company.save();
     return savedCompany;
@@ -136,21 +162,25 @@ async function deleteFromExAssociatedCompany(note) {
   }
 }
 
-async function addToAssociatedPerson(personId, companyId, note) {
+async function addToAssociatedPerson(personId, companyId, note, userId) {
   try {
     const person = await Person.findById(personId);
     if (!person) {
       throw new Error('unable to find person');
     }
+
+    if (userId !== person.author.toString()) {
+      throw new Error('No permission error');
+    }
     if (companyId) {
       if (person.associatedCompany && person.associatedCompany.toString() !== companyId) {
         throw new Error('mismatch between associated company and associated person');
       } else {
-        addToAssociatedCompany(companyId, note);
+        addToAssociatedCompany(companyId, note, userId);
         note.associatedCompany = companyId;
       }
     } else if (person.associatedCompany) {
-      addToAssociatedCompany(person.associatedCompany, note);
+      addToAssociatedCompany(person.associatedCompany, note, userId);
       note.associatedCompany = person.associatedCompany;
     }
     person.notes.push(note.id);
@@ -161,9 +191,16 @@ async function addToAssociatedPerson(personId, companyId, note) {
   }
 }
 
-async function deleteFromExAssociatedPerson(note) {
+async function deleteFromExAssociatedPerson(note, userId) {
   try {
     const person = await Person.findById(note.associatedPerson);
+    if (!person) {
+      throw new Error('unable to find person');
+    }
+
+    if (userId !== person.author.toString()) {
+      throw new Error('No permission error');
+    }
     person.notes.pull(note.id);
     const savedPerson = await person.save();
     return savedPerson;
